pierdoły
git-svn-id: svn+ssh://kadu.net/home/kadu/repos/kadu/trunk@3102 fea5205d-6df5-0310-bef7-ea67e47da9a5

+++ kadu/chat.cpp
@@ -1022,7 +1022,7 @@ void Chat::closeEvent(QCloseEvent* e)
 
 	if (config_file.readBoolEntry("Chat", "ChatCloseTimer"))
 	{
-		unsigned long period = config_file.readUnsignedNumEntry("Chat",
+		unsigned int period = config_file.readUnsignedNumEntry("Chat",
 			"ChatCloseTimerPeriod", 2);
 		if (QDateTime::currentDateTime() < lastMsgTime.addSecs(period))
 		{
